# Fix issue #71: Optimize nested structure decoding performance