@Jecoms 2025-10-04
I'd like to further improve the performance, but this is a start.